+++ jason/flutter/lib/reconnect_handle.dart
@@ -1,5 +1,6 @@
 import 'dart:ffi';
 
+import 'ffi/foreign_value.dart';
 import 'jason.dart';
 import 'util/move_semantic.dart';
 import 'util/nullable_pointer.dart';
@@ -11,9 +12,9 @@ typedef _reconnect_with_delay_C = Handle Function(Pointer, Int64);
 typedef _reconnect_with_delay_Dart = Object Function(Pointer, int);
 
 typedef _reconnect_with_backoff_C = Handle Function(
-    Pointer, Int64, Double, Int64);
+    Pointer, Int64, Double, Int64, ForeignValue);
 typedef _reconnect_with_backoff_Dart = Object Function(
-    Pointer, int, double, int);
+    Pointer, int, double, int, ForeignValue);
 
 final _free = dl.lookupFunction<_free_C, _free_Dart>('ReconnectHandle__free');
 
@@ -36,7 +37,7 @@ class ReconnectHandle {
   /// provided [Pointer].
   ReconnectHandle(this.ptr);
 
-  // TODO: Add throws docs when all errros are implemented.
+  // TODO: Add throws docs when all errors are implemented.
   /// Tries to reconnect a `Room` after the provided delay in milliseconds.
   ///
   /// If the `Room` is already reconnecting then new reconnection attempt won't
@@ -46,28 +47,35 @@ class ReconnectHandle {
     await (_reconnect_with_delay(ptr.getInnerPtr(), delayMs) as Future);
   }
 
-  // TODO: Add throws docs when all errros are implemented.
+  // TODO: Add throws docs when all errors are implemented.
   /// Tries to reconnect a `Room` in a loop with a growing backoff delay.
   ///
-  /// The first attempt to reconnect is guaranteed to happen not earlier than
-  /// [starting_delay_ms].
+  /// The first attempt will be performed immediately, and the second attempt
+  /// will be performed after [starting_delay_ms].
   ///
-  /// Also, it guarantees that delay between reconnection attempts won't be
-  /// greater than [max_delay_ms].
+  /// Delay between reconnection attempts won't be greater than [max_delay_ms].
   ///
   /// After each reconnection attempt, delay between reconnections will be
   /// multiplied by the given [multiplier] until it reaches [max_delay_ms].
   ///
+  /// If [multiplier] is a negative number then it will be considered as `0.0`.
+  /// This might cause a busy loop, so it's not recommended.
+  ///
+  /// Max elapsed time can be limited with an optional [maxElapsedTimeMs]
+  /// argument.
+  ///
   /// If the `Room` is already reconnecting then new reconnection attempt won't
   /// be performed. Instead, it will wait for the first reconnection attempt
   /// result and use it here.
-  ///
-  /// If [multiplier] is negative number then [multiplier] will be considered as
-  /// `0.0`.
   Future<void> reconnectWithBackoff(
-      int startingDelayMs, double multiplier, int maxDelay) async {
-    await (_reconnect_with_backoff(
-        ptr.getInnerPtr(), startingDelayMs, multiplier, maxDelay) as Future);
+      int startingDelayMs, double multiplier, int maxDelay,
+      [int? maxElapsedTimeMs]) async {
+    var maxElapsedTimeMs_arg = maxElapsedTimeMs == null
+        ? ForeignValue.none()
+        : ForeignValue.fromInt(maxElapsedTimeMs);
+
+    await (_reconnect_with_backoff(ptr.getInnerPtr(), startingDelayMs,
+        multiplier, maxDelay, maxElapsedTimeMs_arg.ref) as Future);
   }
 
   /// Drops the associated Rust struct and nulls the local [Pointer] to it.