+++ tests/e2e/steps/websocket.rs
@@ -12,6 +12,7 @@ async fn ws_connection_loss(world: &mut World, id: String) {
 async fn ws_connection_restore(world: &mut World, id: String) {
     let member = world.get_member(&id).unwrap();
     member.ws_mock().disable_connection_loss().await;
+    member.room().start_ws_reconnect().await.unwrap();
 }
 
 #[then(regex = r"^(\S+)'s WS connection is lost$")]